Docker build secrets: no more API keys in image
Changes the build process to no longer leak the `GITLAB_TOKEN` API key into the `bionemo` Docker image. The `setup/Dockerfile` now expects the `GITLAB_TOKEN` value to be passed-in at build time as a secret instead of a build argument. [1]

As a result, both the `launch.sh` script's `build` command and the `build-bionemo-image` CI step now pass in the env var value for `GITLAB_TOKEN` as a secret. Note that building still allows for specifying this value as an environment variable.

The `internal/Dockerfile-devel` image still has an issue, however. The CI process passes-in `CI_JOB_TOKEN` via the build-arg `TPL_REPO`. [2] This token does expire when the build finishes, so it is a less-critical security leak. Comments with `# FIXME` have been added to this file and the `.gitlab-ci.yml` file to note this problem so it can be fixed in the future.

This implements part one of [CDISCOVERY-2181](https://jirasw.nvidia.com/browse/CDISCOVERY-2181).

[1] https://docs.docker.com/build/building/secrets/

[2] https://docs.gitlab.com/ee/ci/jobs/ci_job_token.html