Fix launch.sh build w/ GITLAB_TOKEN defined in .env
Fixes existing workflow of defining GITLAB_TOKEN in the .env file when
building the image using `launch.sh build`. The `GITLAB_TOKEN` variable
needs to be `export`ed if it is only `source`d from the .env file,
otherwise the `docker build` command will not have access to it in its
environment.