Image Tagging Revamp: CI & launch.sh Changes
No more `latest` tag. Since the `"latest"` image tag name is ambiguous and doesn't refer to a specific version, users must now always `pull` a specific version of `bionemo`. The CI no longer pushes images with the tag `"latest"`.

The paradigm shift in this set of changes is to default to, and thus encourage, using the **`git` commit hash** as the primary means to associate versions of the BioNeMo FW repository to a runnable Docker image environment.


CI Change Details
--
The CI now publishes images using primarily the git commit (SHA) value associated with the build. The build image phase will attempt to `docker pull` the image beforehand, which speeds up repeated pipeline runs on the same (or the previous) commit.

This phase will also tag the image using a the date , the MR number, and a short commit hash when it pushes it to Gitlab's image registry (for merge-request branches only). The format is `mr--RRRR--YYYY-MM-DD--CCCCCCCC`, where `R..R` is the merge request number, and `C..C` is the first 8 characters of the commit hash. Images are still tagged with the prior pipeline tag schema as well. This schema is `pipeline-PPPPPPP`, where `P..P` is the Gitlab CI pipeline ID number. Note that `R` and `P` are variable sized.

The `dev-deploy` stage has been significantly changed. While images are still pushed to both NVCR and Gitlab's image registry, the tagging scheme is completely different. The new tagging scheme includes the branch, date, and short commit hash.

Note that the `-devel` and `-docs` suffixes remain the same for their respective images. In the new tagging schemas, the informative parts are separated with `--`. **NOTE: For `-docs` and `-devel` images using the new tagging schema, one must be aware that this suffix should be stripped or otherwise ignored while parsing the tag.** We recommend using the docker image **labels** to uncover metadata, as this information is immutable and baked-in to the image at build time.



`launch.sh` Change Details
--
New changes to `launch.sh {build, push, dev, run}` commands: they now work with the current git repository commit SHA.
  + `build` and `push` will fail-fast if git is in a dirty state
  + `dev` and `run` will warn, since they volume-mount the local code into the container
  + the image repository name is always parsed from "${BIONEMO_IMAGE}"

**BACKWARDS COMPATIBILITY** To obtain prior `launch.sh` behavior on `build`, `push`, `dev`, and `run`, use the new `-s` (or `--stable`) flag. 

This flag will use `BIONEMO_IMAGE` as the complete image name (image repository + tag) instead.


Tracking
--
Completes [CDISCOVERY-2143](https://jirasw.nvidia.com/browse/CDISCOVERY-2143)